+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-05 16:35:50
-// Commit: fix(state): prevent double tap duplicate action in state
+// Updated at: 2026-07-05 23:11:05
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '2606';
-  final String lastMessage = 'fix(state): prevent double tap duplicate action in state';
-  final DateTime timestamp = DateTime.parse('2026-07-05T16:35:50+05:00');
+  final String commitRef = '2620';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-07-05T23:11:05+05:00');
 }